requirements.txt
# QuantumSentinel-Nexus Comprehensive Security Framework Requirements
# Version 3.0 - Recon, OSINT, and Bug Bounty Tools Integration

# Core Python framework
aiofiles>=23.2.1
aiohttp>=3.9.1
requests>=2.31.0
urllib3>=2.1.0
asyncio-throttle>=1.0.2

# Configuration and data processing
pyyaml>=6.0.1
configparser>=6.0.0
python-dotenv>=1.0.0

# Web scraping and parsing
beautifulsoup4>=4.12.2
lxml>=4.9.3
html5lib>=1.1

# Network and security libraries
dnspython>=2.4.2
python-whois>=0.8.0
cryptography>=41.0.7
pycryptodome>=3.19.0

# OSINT and intelligence gathering
shodan>=1.31.0
censys>=2.2.0
python-twitter>=3.5

# Data analysis and visualization
pandas>=2.1.4
numpy>=1.24.4
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0

# Report generation
jinja2>=3.1.2
weasyprint>=60.1
reportlab>=4.0.7

# PDF and document processing
pypdf2>=3.0.1
pdfkit>=1.0.0
markdown>=3.5.1

# Image processing for screenshots
pillow>=10.1.0

# Database support
sqlalchemy>=2.0.23

# HTTP and web testing
httpx>=0.25.2
flask>=3.0.0
fastapi>=0.104.1
uvicorn>=0.24.0

# Security testing libraries
python-nmap>=0.7.1
scapy>=2.5.0
paramiko>=3.4.0

# Machine learning for AI validation (lightweight versions)
scikit-learn>=1.3.2
textblob>=0.17.1

# Text processing and analysis
nltk>=3.8.1
regex>=2023.10.3

# Time and date handling
python-dateutil>=2.8.2
pytz>=2023.3

# Logging and monitoring
loguru>=0.7.2
colorlog>=6.8.0
rich>=13.7.0
colorama>=0.4.6

# Utility libraries
click>=8.1.7
tqdm>=4.66.1
tabulate>=0.9.0
termcolor>=2.3.0

# File handling and compression
patoolib>=1.12.0

# Git and version control
gitpython>=3.1.40

# Network scanning and analysis
netaddr>=0.10.1

# Web automation (for advanced testing)
selenium>=4.15.2

# API testing and validation
jsonschema>=4.19.2

# Performance and optimization
psutil>=5.9.6

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0

# CLI and interface
argparse
prompt-toolkit>=3.0.41

# Docker and containerization support
docker>=6.1.3

# Mobile Security Framework - Full Integration
# Comprehensive Mobile Security Testing Suite
frida-tools>=12.2.1
objection>=1.11.0
androguard>=3.4.0a1

# 3rd-EAI AI Validation Engine Dependencies
tensorflow>=2.13.0
torch>=2.0.1
transformers>=4.35.0
sklearn>=0.0.post10
xgboost>=2.0.0

# Advanced ML libraries for AI validation
lightgbm>=4.1.0
catboost>=1.2.2

# Video PoC Recording System
opencv-python>=4.8.1.78
moviepy>=1.0.3
imageio>=2.31.5

# iOS Testing Environment
plistlib3>=0.1.0

# Android Testing Environment
python-adb>=3.2.0

# Advanced Exploitation Framework
frida>=16.1.4
frida-tools>=12.2.1

# Additional mobile security tools
pymobiledevice3>=2.15.0

# Enhanced PDF generation for mobile reports
weasyprint>=60.1
reportlab>=4.0.7

# Additional security tools integration
python-whois>=0.8.0
shodan>=1.31.0
censys>=2.2.0

# QuantumSentinel-Nexus v4.0 AI/ML Dependencies
# Core AI/ML Framework
torch>=2.0.1
transformers>=4.35.0
torch-geometric>=2.4.0
scikit-learn>=1.3.2

# Advanced ML Libraries
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.2

# Graph Neural Networks
networkx>=3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0

# Binary Analysis Libraries (optional - install if needed)
# angr>=9.2.0
# capstone>=5.0.0
# keystone-engine>=0.9.2
# pwntools>=4.11.0

# Emulation and Dynamic Analysis (optional)
# qiling>=1.4.0
# unicorn>=2.0.0

# Web Testing Enhancement
aiohttp>=3.9.1
selenium>=4.15.2

# Data Processing
pandas>=2.1.4
numpy>=1.24.4

# Visualization for AI insights
plotly>=5.17.0
seaborn>=0.13.0

# Async processing
asyncio-throttle>=1.0.2

# Additional ML utilities
joblib>=1.3.0
tqdm>=4.66.1